Se decidio que se creara la GUI en AppDesigner de MatLab, ya que es el mas adecuado para el entorno donde esta enfocado el proyecto. Esta primer version del GUi en MatLab, se puede mediante botones iniciar la transmision y detenerla ademas de exportar un archivo tipo excel

+++ CodeController/CodeController.ino
@@ -16,7 +16,7 @@ void loop() {
   digitalWrite(TRIGGER_PIN, LOW);
 
   long duration = pulseIn(ECHO_PIN, HIGH);
-  float distance = (duration * 0.034) / 2;
+  int distance = (duration * 0.034) / 2;
 
   // Imprime la distancia medida
   Serial.println(distance);
